# doc/swagger/true.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Peerbill User",
    "version": "1.0",
    "contact": {
      "name": "Peerbill user",
      "email": "peerbillofficial@gmail.com"
    }
  },
  "tags": [
    {
      "name": "PeerbillUser"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/buy-airtime": {
      "post": {
        "summary": "Purchase airtime",
        "description": "Use this API to purchase airtime",
        "operationId": "PeerbillUser_BuyAirtime",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbAirtimeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbAirtimeRequest"
            }
          }
        ],
        "tags": [
          "PeerbillUser"
        ]
      }
    },
    "/api/gen-wallet": {
      "post": {
        "summary": "Generate escrow wallet",
        "description": "Use this API to generate escrow wallet",
        "operationId": "PeerbillUser_GenWallet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbGenWalletResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbGenWalletRequest"
            }
          }
        ],
        "tags": [
          "PeerbillUser"
        ]
      }
    },
    "/api/get-dataplans": {
      "post": {
        "summary": "Get Data Plans for specified network",
        "description": "Use this API to get Data Plans for specified network",
        "operationId": "PeerbillUser_GetDataPlans",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDataLookupResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbDataLookupRequest"
            }
          }
        ],
        "tags": [
          "PeerbillUser"
        ]
      }
    },
    "/api/get-rates": {
      "post": {
        "summary": "Get Crypto/Fiat rates",
        "description": "Use this API to get Crypto/Fiat rates",
        "operationId": "PeerbillUser_GetRates",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbRateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbRateRequest"
            }
          }
        ],
        "tags": [
          "PeerbillUser"
        ]
      }
    },
    "/api/made-payment": {
      "post": {
        "summary": "Complete trade orders",
        "description": "Use this API to complete trade orders",
        "operationId": "PeerbillUser_MadePayment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbMadePaymentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbMadePaymentRequest"
            }
          }
        ],
        "tags": [
          "PeerbillUser"
        ]
      }
    }
  },
  "definitions": {
    "pbAirtimeRequest": {
      "type": "object",
      "properties": {
        "network": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      }
    },
    "pbAirtimeResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "string"
        }
      }
    },
    "pbDataLookupRequest": {
      "type": "object",
      "properties": {
        "network": {
          "type": "string"
        }
      }
    },
    "pbDataLookupResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean"
        },
        "description": {
          "type": "string"
        },
        "message": {
          "$ref": "#/definitions/pbMessageDetails"
        },
        "statusCode": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "pbGenWalletRequest": {
      "type": "object",
      "properties": {
        "crypto": {
          "type": "string"
        }
      }
    },
    "pbGenWalletResponse": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        }
      }
    },
    "pbMadePaymentRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "username": {
          "type": "string"
        },
        "orderType": {
          "type": "string"
        },
        "userAddress": {
          "type": "string"
        },
        "escrowAddress": {
          "type": "string"
        }
      }
    },
    "pbMadePaymentResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "pbMessageDetails": {
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/pbNetworkDetails"
          }
        }
      }
    },
    "pbNetworkDetails": {
      "type": "object",
      "properties": {
        "networkName": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "networkCode": {
          "type": "string"
        },
        "checkBalance": {
          "type": "string"
        },
        "logoUrl": {
          "type": "string"
        },
        "plans": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/pbPlan"
          }
        }
      }
    },
    "pbPlan": {
      "type": "object",
      "properties": {
        "planCode": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "alias": {
          "type": "string"
        },
        "amount": {
          "type": "string"
        }
      }
    },
    "pbRateRequest": {
      "type": "object",
      "properties": {
        "fiat": {
          "type": "string"
        },
        "crypto": {
          "type": "string"
        }
      }
    },
    "pbRateResponse": {
      "type": "object",
      "properties": {
        "rate": {
          "type": "number",
          "format": "float"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}